Allow comma and spaces in zypper list
Simplify the installation of package defitions for combustion by
allowing users to insert multiple packages, separated either by space or
by comma.

This should make it easier to users to select multiple packages, because
the text field does not require the knowledge that the package list will
be directly used in zypper, but it allows a more human-natural way of
entering comma separated items.